Add kept-rate benchmarks
