+++ crates/edit_prediction_cli/src/lib.rs
@@ -1,0 +1,4 @@
+#[allow(dead_code)]
+mod word_diff;
+
+pub mod kept_rate;
